Docker setup (#54)
* Added Dockerfile and updated README.md accordingly

The Dockerfile is a minimal setup to construct a  layered docker image, which automatically
launches into the nix-shell when executed interactively.

The README.md now contains basic instructions about how to build and run the image, although
it does not yet specify how to e.g. sync the database files to be persistent or other niceties.