Initial Tarball hosting
This commit adds initial support for hosting package tarballs, gated by
the flag `FLORA_TARBALLS_ENABLED`.

This supports reading in tarballs, constructing a directory tree and
storing files and directories by hash. The directory tree relations are
stored in the postgresql database, and the blobs themselves are stored
on the filesystem (or in memory for tests), the directory configured by
`FLORA_TARBALLS_FS_PATH`. This should be expanded to other storage
options in future (e.g. S3).

This method both reduces space of tarballs and removes the need for
metadata revisions, because small updates are cheap.

Tarballs can either be imported by the fetch metadata handler on the
admin page, or individually by the new `import-package-tarball` command
in flora-cli.