Transitive dependencies
== Transitive dependencies API
This PR adds an API endpoint for transitive dependencies.
It relies on a recursive CTE on the PostgreSQL side.

The endpoint takes a query parameter (`?transitive=true`)
to enable the full tree of dependencies.

== Package import concurrency

It so happens that I found a concurrency problem and fixed it with the
great help of @elland. The provisional diagnostic is that Poolboy's
queue
workers were being killed by the GHC as the function that called
`Poolboy.enqueue` then returned, before the completion of the work
queue. Moreover, `Poolboy.waitingStopFinishWorkers` was not behaving
as expected with the blocking.

Poolboy was thus replaced by a call to `pooledForConcurrentlyN_`, which
has a
blocking behaviour.

Fixes #390